% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_to_pseudobulk_modified}
\alias{convert_to_pseudobulk_modified}
\title{Convert Single-cell Count Matrix to Pseudobulk Matrix for a Individual Cell Type}
\usage{
convert_to_pseudobulk_modified(dds, sample_accessor, design = NULL)
}
\arguments{
\item{dds}{A `DESeqDataSet` object containing single-cell count data for an individual cell type.}

\item{sample_accessor}{A character string specifying the column name in `colData(dds)`
that contains sample labels.}

\item{design}{A formula or matrix specifying the experimental design.
The formula defines how the counts for each gene depend on the variables
in `colData(dds)`. Default is `~1` (intercept only).}
}
\value{
A `DESeqDataSet` object with pseudobulk-level counts aggregated by sample.
}
\description{
This function aggregates single-cell count data into pseudobulk count data
at the sample level for an individual cell type. It generates a `DESeqDataSet`
object at the pseudobulk level, which can be used for downstream differential
expression analyses.
}
