% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_subclusters}
\alias{create_subclusters}
\title{Create Subclusters by Cell Type}
\usage{
create_subclusters(dds, accessor)
}
\arguments{
\item{dds}{A `DESeqDataSet` object of data for all cell types.}

\item{accessor}{A character string specifying the column name in `colData(dds)`
that identifies cell types.}
}
\value{
A named list of `DESeqDataSet` objects, where each element corresponds
to a single cell type and contains only the cells of that type.
}
\description{
This function splits a DESeqDataSet object into subclusters based on a
specified cell type accessor. Each subcluster contains data for a single cell
type, facilitating downstream analysis specific to each cell type.
}
