% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter Genes by Expression Threshold for an Individual Cell Type}
\usage{
filter_genes(dds, pct = 0.15)
}
\arguments{
\item{dds}{A `DESeqDataSet` object representing cells from an individual cell type.}

\item{pct}{A numeric value (between 0 and 1) specifying the minimum percentage of
cells in which a gene must be expressed to be retained. Default is 15%.}
}
\value{
A filtered `DESeqDataSet` object containing only genes that meet the
   specified expression threshold.
}
\description{
This function filters genes in a DESeqDataSet based on the percentage of cells
in which they are expressed. It retains only those genes whose expression exceeds
a specified threshold, facilitating downstream analyses with more relevant genes.
}
