% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_low_celltypes}
\alias{filter_low_celltypes}
\title{Filter Samples Based on Cell Counts for an Individual Cell Type}
\usage{
filter_low_celltypes(dds, sample, threshold = 50)
}
\arguments{
\item{dds}{A `DESeqDataSet` object representing cells from an individual cell type.}

\item{sample}{A character string specifying the column name in `colData(dds)`
that contains sample labels.}

\item{threshold}{An integer specifying the minimum number of cells required
for a sample to be retained. Default is 50.}
}
\value{
A filtered `DESeqDataSet` object containing only samples with
   sufficient cell counts.
}
\description{
This function filters samples in a DESeqDataSet based on the number of cells
for a specified cell type. It retains only samples with cell counts exceeding
a given threshold, facilitating analyses focused on well-represented samples.
}
