% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_PseuLME}
\alias{fit_PseuLME}
\title{Cell Type Specific Pseudobulk Differential Expression Analysis via Linear Mixed-Effects Model}
\usage{
fit_PseuLME(
  dds,
  cell_type_accessor,
  condition_accessor,
  sample_accessor,
  batch_accessor,
  gene_pct,
  sample_thres,
  contrast1,
  contrast2
)
}
\arguments{
\item{dds}{A `DESeqDataSet` object containing single-cell or pseudobulk count data.}

\item{cell_type_accessor}{A character string specifying the column name in `colData(dds)` that identifies cell types.}

\item{condition_accessor}{A character string specifying the column name in `colData(dds)` that identifies condition labels (e.g., treatment vs. control).}

\item{sample_accessor}{A character string specifying the column name in `colData(dds)` that identifies sample labels.}

\item{batch_accessor}{A character string specifying the column name in `colData(dds)` that identifies batch labels.}

\item{gene_pct}{A numeric value (between 0 and 1) indicating the minimum percentage of cells in which a gene must be expressed to be retained.}

\item{sample_thres}{An integer specifying the minimum number of cells required for a sample to be retained in the analysis.}

\item{contrast1}{A character string representing the label of the control group in the condition column.}

\item{contrast2}{A character string representing the label of the case group in the condition column.}
}
\value{
A named list of data frames, where each element corresponds to a cell type subcluster.
   Each data frame contains the following columns for each gene:
  - `base`: The baseline expression.
  - `log2fc`: The log2 fold change between `contrast2` and `contrast1`.
  - `pvals`: The p-value for the differential expression test.
  - `padjs`: The FDR-adjusted p-value (using the Benjamini-Hochberg method).
}
\description{
This function performs differential expression analysis for subclusters of single-cell data by:
1. Creating subclusters based on a specified cell type.
2. Filtering genes and samples based on user-defined thresholds.
3. Aggregating single-cell data into pseudobulk counts for each sample.
4. Fitting linear mixed-effects models to estimate differential expression between two conditions.
}
