% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lme_pvals}
\alias{lme_pvals}
\title{Fit Linear Mixed-Effect Model on Normalized Gene Expression Data}
\usage{
lme_pvals(dds, condition_accessor, batch_accessor, contrast1, contrast2)
}
\arguments{
\item{dds}{A `DESeqDataSet` object containing pseudobulk count data for an individual cell type.}

\item{condition_accessor}{A character string specifying the column name in
`colData(dds)` that contains condition labels (e.g., treatment vs. control).}

\item{batch_accessor}{A character string specifying the column name in
`colData(dds)` that contains batch labels (e.g., assay 1,2,3,...).}

\item{contrast1}{A character string representing the label of the control group.}

\item{contrast2}{A character string representing the label of the case group.}
}
\value{
A data frame containing the following columns:
  - `base`: The baseline expression for each gene.
  - `log2fc`: The log2 fold change between `contrast2` and `contrast1`.
  - `pvals`: The p-value for the differential expression analysis.
  - `padjs`: The FDR-adjusted p-value (using the Benjamini-Hochberg method).
  The rows of the data frame correspond to genes in the `DESeqDataSet`.
}
\description{
Fit Linear Mixed-Effect Model on Normalized Gene Expression Data
}
