% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_pseudobulk}
\alias{plot_pseudobulk}
\title{Plot Pseudobulk Gene Expression}
\usage{
plot_pseudobulk(dds, contrast1, contrast2, gene, log2 = FALSE)
}
\arguments{
\item{dds}{A `DESeqDataSet` object containing pseudobulk count matrix of a specific cell type.}

\item{contrast1}{A character string representing the control group label in the condition column.}

\item{contrast2}{A character string representing the case group label in the condition column.}

\item{gene}{A character string indicating the gene name of interest.}

\item{log2}{A logical value showing whether to log2-transform the expression values (default: FALSE).}
}
\value{
A ggplot object showing psuedobulk gene expression between 2 conditions.
}
\description{
Generate a plot of pseudobulk gene expression between 2 conditions.
}
