% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_pseudobulk}
\alias{plot_pseudobulk}
\title{Draw Pseudobulk Expression}
\usage{
plot_pseudobulk(dds, contrast1, contrast2, gene, log2 = FALSE)
}
\arguments{
\item{dds}{A `DESeqDataSet` object containing single-cell or pseudobulk count data.}

\item{contrast1}{A character string representing the label of the control group in the condition column.}

\item{contrast2}{A character string representing the label of the case group in the condition column.}

\item{gene}{A character string of the gene name of interest.}

\item{log2}{A binary value showing whether the expression is log2-transformed.}
}
\value{
ggplot of psuedobulk expression
}
\description{
Draw Pseudobulk Expression
}
